pyproject.toml
[project]
name = "Practical_Python_Code_Adventure"
version = "0.2.0"
description = "Yearly coding challenges for the Practical Python Discord Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.5",
    "discord-py~=2.3.2",
    "flask~=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv~=1.0.1",
    "pyyaml~=6.0",
    "requests~=2.31.0",
    "sqlalchemy~=2.0.31",
]
